# Add support for building optimized string table