@xoofx 2023-10-03
Thanks! Results look great, definitely agree that the current implementation of `ElfStringTable` is super inefficient.

Will have a look at the PR later tonight.