@xoofx 2023-10-05
Looks like the CI was stuck because it was requesting `ubuntu-18.04` but it has been removed from GitHub action.

I have switched to `ubuntu-20.04` but I expect tests to fail as they were relying on a particular version of readelf.

Also, when running locally, I got multiple errors from the DwarfTests, and they seem to be related to the string table changes. For example:

```
System.IndexOutOfRangeException : Index was outside the bounds of the array.
   at LibObjectFile.ObjectFileStreamExtensions.ReadStringUTF8NullTerminated(Stream stream) in C:\code\LibObjectFile\src\LibObjectFile\ObjectFileStreamExtensions.cs:line 71
   at LibObjectFile.Dwarf.DwarfStringTable.GetStringFromOffset(UInt64 offset) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfStringTable.cs:line 43
   at LibObjectFile.Dwarf.DwarfAttribute.ReadAttributeFormRawValue(DwarfReader reader) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfAttribute.cs:line 435
   at LibObjectFile.Dwarf.DwarfAttribute.Read(DwarfReader reader) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfAttribute.cs:line 283
   at LibObjectFile.Dwarf.DwarfObject`1.ReadInternal(DwarfReader reader) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfObject.cs:line 79
   at LibObjectFile.Dwarf.DwarfDIE.Read(DwarfReader reader) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfDIE.cs:line 415
   at LibObjectFile.Dwarf.DwarfObject`1.ReadInternal(DwarfReader reader) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfObject.cs:line 79
   at LibObjectFile.Dwarf.DwarfDIE.ReadInstance(DwarfReader reader) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfDIE.cs:line 457
   at LibObjectFile.Dwarf.DwarfUnit.Read(DwarfReader reader) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfUnit.cs:line 151
   at LibObjectFile.Dwarf.DwarfObject`1.ReadInternal(DwarfReader reader) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfObject.cs:line 79
   at LibObjectFile.Dwarf.DwarfUnit.ReadInstance(DwarfReader reader, UInt64& offsetEndOfUnit) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfUnit.cs:line 206
   at LibObjectFile.Dwarf.DwarfInfoSection.Read(DwarfReader reader) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfInfoSection.cs:line 49
   at LibObjectFile.Dwarf.DwarfObject`1.ReadInternal(DwarfReader reader) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfObject.cs:line 79
   at LibObjectFile.Dwarf.DwarfFile.Read(DwarfReaderContext readerContext) in C:\code\LibObjectFile\src\LibObjectFile\Dwarf\DwarfFile.cs:line 421
   at LibObjectFile.Tests.Dwarf.DwarfTests.TestDebugInfoSmall() in C:\code\LibObjectFile\src\LibObjectFile.Tests\Dwarf\DwarfTests.cs:line 328
   at System.RuntimeMethodHandle.InvokeMethod(Object target, Void** arguments, Signature sig, Boolean isConstructor)
   at System.Reflection.MethodInvoker.Invoke(Object obj, IntPtr* args, BindingFlags invokeAttr)
```