@filipnavara 2023-10-05
That particular error is fixed with https://github.com/xoofx/LibObjectFile/pull/27.

I also had trouble running the tests locally for the reasons you mentioned.