@xoofx 2023-10-05
> That particular error is fixed with https://github.com/xoofx/LibObjectFile/pull/27.

Oh, good catch, just merged!

I'm starting to ask myself If I should not make this library only compatible with net7.0.

Would it be a problem for your experiment?