@filipnavara 2023-10-05
> I'm starting to ask myself If I should not make this library only compatible with net7.0.
> 
> Would it be a problem for your experiment?

Not at all. I am fine with that. I think I publish my Mach-O libraries as net6.0+ since that's still supported LTS release but I don't mind a higher version either.