@xoofx 2023-10-05
Ok, just bumped on master to net7.0.

`TestHelloWorld` is failing. As expected there are new stuffs in `ubuntu-20.04` generated by the gcc version that LibObjectFile is not able to decode.

Might ignore it for a time and add an issue. Don't have time to investigate this one and it might take some time to upgrade.