@filipnavara 2023-10-05
> `TestHelloWorld` is failing. As expected there are new stuffs in `ubuntu-20.04` generated by the gcc version that LibObjectFile is not able to decode.

I may get to look into it, but this week is exceptionally busy, so I cannot promise any timeline.